add debug part
	1. change print => jprintf
	2. add print color(jiprintf, jeprintf)

+++ Inc/firmware_info.h
@@ -38,7 +38,7 @@ typedef struct _SAppInfo SAppInfo;
 
 struct _SFwInfo
 {
-	uint32_t	mucUpdated;									// update 진행 유무
+	uint8_t		mucUpdated;									// update 진행 유무
 	uint8_t		marrucModelName[MODEL_NAME_SIZE];			// Model Name
 	uint8_t		marrucSerialNo[SERIAL_NUMBER_SIZE];			// Serial Number
 	SAppInfo	mstruBlInfo;								// Bootloader Info
